+++ src/test/groovy/nebula/plugin/publishing/maven/MavenManifestPluginIntegrationSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Netflix, Inc.
+ * Copyright 2015-2017 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,15 @@
  */
 package nebula.plugin.publishing.maven
 
-import nebula.test.IntegrationSpec
+import nebula.test.IntegrationTestKitSpec
 
-class MavenManifestPluginIntegrationSpec extends IntegrationSpec {
+class MavenManifestPluginIntegrationSpec extends IntegrationTestKitSpec {
     def setup() {
         buildFile << """\
-            apply plugin: 'nebula.maven-manifest'
-            apply plugin: 'nebula.maven-nebula-publish'
+            plugins {
+                id 'nebula.maven-manifest'
+                id 'nebula.maven-nebula-publish'
+            }
 
             version = '0.1.0'
             group = 'test.nebula'
@@ -39,7 +41,7 @@ class MavenManifestPluginIntegrationSpec extends IntegrationSpec {
         '''
 
         when:
-        runTasksSuccessfully('generatePomFileForNebulaPublication')
+        runTasks('generatePomFileForNebulaPublication')
 
         then:
         def pom = new XmlSlurper().parse(new File(projectDir, 'build/publications/nebula/pom-default.xml'))