+++ src/test/groovy/nebula/plugin/publishing/maven/MavenResolvedDependenciesPluginIntegrationSpec.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Netflix, Inc.
+ * Copyright 2015-2017 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,18 +15,20 @@
  */
 package nebula.plugin.publishing.maven
 
-import nebula.test.IntegrationSpec
+import nebula.test.IntegrationTestKitSpec
 import nebula.test.dependencies.DependencyGraphBuilder
 import nebula.test.dependencies.GradleDependencyGenerator
 import nebula.test.dependencies.ModuleBuilder
 
-class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
+class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationTestKitSpec {
     File publishDir
 
     def setup() {
         buildFile << """\
-            apply plugin: 'nebula.maven-resolved-dependencies'
-            apply plugin: 'nebula.maven-nebula-publish'
+            plugins {
+                id 'nebula.maven-resolved-dependencies'
+                id 'nebula.maven-nebula-publish'
+            }
 
             version = '0.1.0'
             group = 'test.nebula'
@@ -64,7 +66,30 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
         """.stripIndent()
 
         when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
+        runTasks('publishNebulaPublicationToTestLocalRepository')
+
+        then:
+        def a = findDependency('a')
+        a.version.text() == '1.1.0'
+    }
+
+    def 'dynamic latest versions are replaced by the resolved version'() {
+        def graph = new DependencyGraphBuilder().addModule('test.resolved:a:1.0.0')
+                .addModule('test.resolved:a:1.1.0').build()
+        File mavenrepo = new GradleDependencyGenerator(graph, "${projectDir}/testrepogen").generateTestMavenRepo()
+
+        buildFile << """\
+            apply plugin: 'java'
+
+            repositories { maven { url '${mavenrepo.absolutePath}' } }
+
+            dependencies {
+                compile 'test.resolved:a:latest.release'
+            }
+        """.stripIndent()
+
+        when:
+        runTasks('publishNebulaPublicationToTestLocalRepository')
 
         then:
         def a = findDependency('a')
@@ -87,7 +112,7 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
         """.stripIndent()
 
         when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
+        runTasks('publishNebulaPublicationToTestLocalRepository')
 
         then:
         def d = findDependency('d')
@@ -112,7 +137,7 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
         """.stripIndent()
 
         when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
+        runTasks('publishNebulaPublicationToTestLocalRepository')
 
         then:
         def a = findDependency('a')
@@ -125,10 +150,26 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
                 .build()
         File mavenrepo = new GradleDependencyGenerator(graph, "${projectDir}/testrepogen").generateTestMavenRepo()
 
-        buildFile << """
+        buildFile.text = """\
+            plugins {
+                id 'nebula.maven-resolved-dependencies'
+                id 'nebula.maven-nebula-publish'
+            }
             allprojects {
                 apply plugin: 'java'
-                ${applyPlugin(MavenResolvedDependenciesPlugin)}
+                apply plugin: 'nebula.maven-resolved-dependencies'
+                
+                version = '0.1.0'
+                group = 'test.nebula'
+    
+                publishing {
+                    repositories {
+                        maven {
+                            name = 'testLocal'
+                            url = 'testrepo'
+                        }
+                    }
+                }
 
                 repositories { maven { url '${mavenrepo.absolutePath}' } }
             }
@@ -136,7 +177,7 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
             dependencies {
                 compile project(':sub')
             }
-        """.stripIndent()
+            """.stripIndent()
 
         addSubproject('sub', '''\
             group = 'nebula.hello'
@@ -144,7 +185,7 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
         '''.stripIndent())
 
         when:
-        runTasksSuccessfully('publishNebulaPublicationToTestLocalRepository')
+        runTasks('publishNebulaPublicationToTestLocalRepository')
 
         then:
         def b = findDependency('sub')
@@ -199,14 +240,25 @@ class MavenResolvedDependenciesPluginIntegrationSpec extends IntegrationSpec {
     }
 
     def 'works with java-library plugin and dependency-recommender'() {
-        buildFile << '''\
-            buildscript {
-                repositories { jcenter() }
-                dependencies { classpath 'com.netflix.nebula:nebula-dependency-recommender:4.1.0' }
+        buildFile.text = '''\
+            plugins {
+                id 'nebula.maven-resolved-dependencies'
+                id 'nebula.maven-nebula-publish'
+                id 'java-library'
+                id 'nebula.dependency-recommender' version '4.1.0'
             }
 
-            apply plugin: 'java-library'
-            apply plugin: 'nebula.dependency-recommender'            
+            version = '0.1.0'
+            group = 'test.nebula'
+
+            publishing {
+                repositories {
+                    maven {
+                        name = 'testLocal'
+                        url = 'testrepo'
+                    }
+                }
+            }          
             
             repositories { jcenter() }
             dependencies {