+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+5.0.0-rc.1 / 2017-05-23
+=======================
+
+* Build with gradle 4.0-milestone-2 as we saw issues using our plugin built with 3.5 on a 4.0-milestone-2 test project
+* BREAKING: removed deprecated plugins: `nebula.test-jar` and old names `nebula-test-jar`, `nebula-source-jar`, `nebula-javadoc-jar`
+
 4.9.1 / 2016-08-16
 ==================